requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
decorator==5.1.0
Django==3.2.8
django-appconf==1.0.5
django-imagekit==4.0.2
django-rest-knox==4.1.0
djangorestframework==3.12.4
dynaconf==3.1.7
hvac==0.11.2
idna==3.2
loguru==0.5.3
pilkit==2.0
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
pytz==2021.3
requests==2.26.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
validators==0.18.2
win32-setctime==1.0.3
